# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anophel API Documentation",
        "version": "0.0.0"
    },
    "paths": {
...
        "/api/expertise/{id}": {
            "get": {
                "tags": [
                    "Expertises"
                ],
                "summary": "Get Expertise by ID",
                "description": "Retrieve an expertise by its ID",
                "operationId": "4e9b69529b80ee2055f9d9f766c55e7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Expertise ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "expertise": {
                                                    "properties": {
                                                        "id": {
                                                            "example": "1"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Expertise Name"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Expertises"
                ],
                "summary": "Delete Expertise by ID",
                "description": "Delete an expertise by its ID",
                "operationId": "570cb57b37a44b8a956d84f5e344eed9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Expertise ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "string",
                                            "example": "تخصص مورد نظر با موفقیت حذف شد."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
}